ask before overwrite config files (fixes #17)
added shell module to work with files (check if exists and ask the user for overwrite actions)

fixes #17

+++ lib/capistrano/tasks/logrotate.rake
@@ -1,10 +1,13 @@
 namespace :logrotate do
+  include Shell
 
   desc "Install the logrotate config"
   task :setup do
     on roles(:app) do
-      template "logrotate.erb", "/tmp/logrotate-#{fetch(:application)}"
-      sudo "mv /tmp/logrotate-#{fetch(:application)} /etc/logrotate.d/#{fetch(:application)}"
+      if file_new_or_overwrite?("/etc/logrotate.d/#{fetch(:application)}")
+        template "logrotate.erb", "/tmp/logrotate-#{fetch(:application)}"
+        sudo "mv /tmp/logrotate-#{fetch(:application)} /etc/logrotate.d/#{fetch(:application)}"
+      end
       sudo "chown root:root /etc/logrotate.d/#{fetch(:application)}"
     end
   end
